CAS-11616.  If VLA data are pre-averaged, then hanningsmooth() is not executed.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@41735 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ pipeline/pipeline/hifv/tasks/hanning/hanning.py
@@ -7,6 +7,7 @@ import pipeline.infrastructure as infrastructure
 import pipeline.infrastructure.basetask as basetask
 import pipeline.infrastructure.vdp as vdp
 from pipeline.infrastructure import casa_tasks
+import pipeline.infrastructure.casatools as casatools
 from pipeline.infrastructure import task_registry
 
 LOG = infrastructure.get_logger(__name__)
@@ -47,15 +48,18 @@ class Hanning(basetask.StandardTaskTemplate):
     
     def prepare(self):
 
-        if not self._executor._dry_run:
-            try:
-                self._do_hanningsmooth()
-                LOG.info("Removing original VIS " + self.inputs.vis)
-                shutil.rmtree(self.inputs.vis)
-                LOG.info("Renaming temphanning.ms to " + self.inputs.vis)
-                os.rename('temphanning.ms', self.inputs.vis)
-            except Exception as ex:
-                LOG.warn('Problem encountered with hanning smoothing. ' + str(ex))
+        if self._checkpreaveraged():
+            if not self._executor._dry_run:
+                try:
+                    self._do_hanningsmooth()
+                    LOG.info("Removing original VIS " + self.inputs.vis)
+                    shutil.rmtree(self.inputs.vis)
+                    LOG.info("Renaming temphanning.ms to " + self.inputs.vis)
+                    os.rename('temphanning.ms', self.inputs.vis)
+                except Exception as ex:
+                    LOG.warn('Problem encountered with hanning smoothing. ' + str(ex))
+        else:
+            LOG.warn("Data in this MS are pre-averaged.  CASA task hanningsmooth() was not executed.")
 
         return HanningResults()
     
@@ -69,3 +73,12 @@ class Hanning(basetask.StandardTaskTemplate):
                                         outputvis='temphanning.ms')
 
         return self._executor.execute(task)
+
+    def _checkpreaveraged(self):
+
+        with casatools.TableReader(self.inputs.vis + '/SPECTRAL_WINDOW') as table:
+            effective_bw = table.getcol('EFFECTIVE_BW')
+            resolution = table.getcol('RESOLUTION')
+
+        return not(resolution[0][0] < effective_bw[0][0])
+
